requirements.txt
streamlit>=1.32.0
pillow>=10.0.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.25.0
google-cloud-vision>=3.4.0
pytesseract>=0.3.10
pdf2image>=1.17.0
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.26.0
